+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.2.1"></a>
+## [1.2.1](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.2.0...@times-components/brightcove-video@1.2.1) (2017-10-03)
+
+
+
+
 <a name="1.2.0"></a>
 # [1.2.0](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.0.5...@times-components/brightcove-video@1.2.0) (2017-09-22)